feat: refactor password reset and forgot password components to utilize shared parseMessage utility; enhance loading state handling
